apps/panaversity-fs-py/pyproject.toml
[project]
name = "panaversity-fs"
version = "0.1.0"
description = "Agent-Native Multi-Book Storage System - Python MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "cryptography>=44.0.0",  # RS256 JWT verification
    "greenlet>=3.2.4",
    "httpx>=0.28.0",  # Async HTTP for JWKS fetching
    "hypothesis>=6.148.6",
    "mcp>=1.22.0",
    "opendal>=0.46.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
    "supabase>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# E402: Module level import not at top of file - needed for scripts that modify sys.path
per-file-ignores = { "scripts/*.py" = ["E402"], "tests/scripts/*.py" = ["E402"] }

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/panaversity_fs"]
